// src/models/user.ts
export class User {
    USERID: string;
    FIRSTNAME: string;
    LASTNAME: string;
    EMAIL: string;
    PHONE: string;
    PID?: string;
    PNAME?: string;
    PASSWORD: string;
    USERNAME:string;
    STATUS:boolean;
    TITLE:string;
    SSN:string;
    LEVID?:string;
    TID:string;
    LEVEL?:string;
    TYPE_NAME:string;
    CWT:string;
    CWT_NAME:string;
}